Подключил админку к столовой. Столовая работает, но картинки на сервере пока не удаляются.

+++ app/Http/Controllers/Api/CanteenController.php
@@ -3,12 +3,14 @@
 namespace App\Http\Controllers\Api;
 
 use App\Http\Controllers\Controller;
+use App\Models\Admins;
 use App\Models\Barter;
 use App\Models\Canteen;
 use App\Models\News;
 use Illuminate\Http\Request;
 
-class CanteenController extends Controller{
+class CanteenController extends Controller
+{
 
     /**
      * Display a listing of the resource.
@@ -22,55 +24,73 @@ class CanteenController extends Controller{
      * Store a newly created resource in storage.
      */
     //	title+	type+	price+	img+
-    public function store(Request $request)
+    public function store(Request $request): \Illuminate\Http\JsonResponse
     {
-        $canteen = new Canteen();
-        $canteen->title = $request->get("title");
-        $canteen->img = $request->get("img");
-        $canteen->setImage($canteen, $request);
-        //////////////////////
-        $canteen->price = $request->get("price");
-        $canteen->type = $request->get("type");
-
-        $canteen->save();
-        return "successful";
+        $user = $request->get("user");
+        if (Admins::where("users", $user)->exists()) {
+            $canteen = new Canteen();
+            $canteen->title = $request->get("title");
+            $canteen->price = $request->get("price");
+            $canteen->type = $request->get("type");
+            $canteen->img = $request->get("img");
+            $canteen->setImage($canteen, $request);
+            $canteen->save();
+            return response()->json(['status' => 'success']);
+        }
+        else {
+            return response()->json(['status' => 'fail']);
+        }
     }
 
     /**
      * Display the specified resource.
      */
-    public function show(string $id)
+    public function show(string $id): \Illuminate\Http\JsonResponse
     {
         $canteen = Canteen::find($id);
-        return "successful";
+        return response()->json(['status' => 'success']);
     }
 
     /**
      * Update the specified resource in storage.
      */
-    public function update(Request $request, string $id)
+    public function update(Request $request, string $id): \Illuminate\Http\JsonResponse
     {
-        $canteen = Canteen::find($id);
-        //return $news;
-        $canteen->title = $request->get("title");
-        $canteen->img = $request->get("img");
-        $canteen->setImage($canteen, $request);
-        //////////////////////
-        $canteen->price = $request->get("price");
-        $canteen->type = $request->get("type");
-
-        $canteen->save();
-        return "successful";
+        $user = $request->get("user");
+        if (Admins::where("users", $user)->exists()) {
+            $canteen = Canteen::find($id);
+            $canteen->title = $request->get("title");
+            $canteen->price = $request->get("price");
+            $canteen->type = $request->get("type");
+            if ($request->get("img") != "unchanged") {
+                // TODO: Удалить старую картинку из storage
+                $canteen->img = $request->get("img");
+                $canteen->setImage($canteen, $request);
+            }
+            $canteen->save();
+            return response()->json(['status' => 'success']);
+        }
+        else {
+            return response()->json(['status' => 'fail']);
+        }
     }
 
     /**
      * Remove the specified resource from storage.
      */
-    public function destroy(string $id)
+    public function destroy(string $id, string $user): \Illuminate\Http\JsonResponse
     {
-        $canteen = Barter::find($id);
-        $canteen->forceDelete();
-        return "successful";
+        if (Admins::where("users", $user)->exists()) {
+            $canteen = Canteen::find($id);
+            if ($canteen) {
+                $canteen->forceDelete();
+                // TODO: Удалить старую картинку из storage
+                return response()->json(['status' => 'success', 'message' => 'Dish successfully deleted']);
+            } else {
+                return response()->json(['status' => 'error', 'message' => 'Dish not found'], 404);
+            }
+        } else {
+            return response()->json(['status' => 'error', 'message' => 'No admin rights'], 401);
+        }
     }
-
 }
